p880: Solve IMEI and mobile data problems. Implement baseband recovery.
I don't like the old RILJ, at all. It is unstable and contains just the
necessary code to make RIL to work.
In the old one, there wasn't any baseband recovery management and no quirks
at all.

EternityProject's version instead, contains baseband recovery, that solves
the ugly and bad problem that happens when the radio goes to UNAVAILABLE
state by restarting it completely, avoiding to lose calls and SMS'es.
Also, it contains rawip insmod/rmmod that, on some modern kernels, like
3.4.37 by EternityProject, isn't working like expected.

It also has a better and proper implementation of
RIL_UNSOL_LGE_SELECTED_SPEECH_CODEC management. The reply has to be
an integer value, so that lge-ril.so will handle that correctly.

Moreover, RIL_UNSOL_LGE_PBREADY was not implemented at all and it was
needed to tell the radio to go into AT READY state.

The reason why we need this is because this will actually allow us to
use modern kernels and to (in future) probably get rid of some proprietary
libraries to switch, instead, to opensource code.

Change-Id: I51032840d115761e51e3df135167e541d0957450